requirements.txt
numpy
pillow
matplotlib
sklearn